feat(ui): add UI interactivity for game status/restarting
Adds interactivity to indicate winning players, Starting and restarting a game, and an overall
cleanup to change the gameplayController to a factory since there can be many games played, and it's
not just a single one. Each game, in my mind, should be independent of each other with the current
implementation.